added fastapi cli for deployment with fly
