+++ api/pyproject.toml
@@ -7,7 +7,7 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.10"
-fastapi = "^0.115.4"
+fastapi = {extras = ["standard"], version = "^0.115.4"}
 uvicorn = "^0.32.0"
 pandas = "^2.2.3"
 ipython = "^8.29.0"
